Created Summary Interface, implementing the "Interface Segregation Principle" of the SOLID principles. Now, we can use alternate implementations of summaries, like Victor's subsum. Also created a summaryFactory.
